+++ exonum-java-binding/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Protobuf-based map proofs, see `ProofMapIndexProxy#getProof` and `MapProof`.
+
 ## [0.9.0-rc1] - 2019-12-TBD
 
 ### Overview